+++ changelog.md
@@ -1,3 +1,6 @@
+# 2021-12-01
+- Add Daily & Rec display to bingo selection
+
 # 2021-11-24
 - Add & adjust some bingo fields
 - Exclude Sunqua normal fields from CM